+++ src/components/Ghost/PostCard/index.tsx
@@ -1,11 +1,10 @@
 import React from 'react'
 import Link from 'next/link'
 import { PostOrPage } from '@tryghost/content-api'
-import Image from 'next/image'
-import ReactPlaceholder from 'react-placeholder'
 
 import { ACCENT_COLOR, WHITE_COLOR } from '../../../theme/color'
 import { isClientStory, isExternalContent, getFirstHref } from '../../../helpers/Ghost/post'
+import ImageWithPlaceholder from '../../ImageWithPlaceholder'
 
 import {
   Anchor,
@@ -54,31 +53,20 @@ const GhostPostCard = (props: GhostPostCardProps) => {
             <CardImage isVisible={post?.feature_image?.length > 0}>
             {post?.feature_image
               ? (
-                <>
-                    <ReactPlaceholder
-                      type='rect'
-                      firstLaunchOnly={false}
-                      showLoadingAnimation={true}
-                      ready={false}
-                      style={{ marginRight: '0', position: 'absolute' }}
-                      // eslint-disable-next-line react/no-children-prop
-                      children={null}
-                    />
-                    <Image
-                      src={post.feature_image}
-                      layout="fill"
-                      objectFit="cover"
-                      aria-label={`post ${post.title}`}
-                      title={`post ${post.title}`}
-                    />
-                  </>
+                  <ImageWithPlaceholder
+                    imageUrl={post.feature_image}
+                    imageAriaLabel={`post ${post.title}`}
+                    imageTitle={`post ${post.title}`}
+                    placeholderType='rect'
+                    style={{ position: 'absolute', top: '0' }}
+                  />
                 )
               : null }
           </CardImage>
             <CardHeadline>{post.title}</CardHeadline>
           </Header>
           <EntryContent>
-            <p>{post.excerpt}</p>
+            <p>{post.custom_excerpt}</p>
             <CardPublishedAt isShowDate={!isClientStory(post)}>{post.published_at}</CardPublishedAt>
           </EntryContent>
         </Article>